requirements/quality.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
aiohttp==3.8.3            # via -r requirements/test.txt, fast-bitrix24
aiosignal==1.2.0          # via -r requirements/test.txt, aiohttp
astroid==2.12.10          # via pylint, pylint-celery
asttokens==2.0.8          # via -r requirements/test.txt, icontract
async-timeout==4.0.2      # via -r requirements/test.txt, aiohttp
asyncio==3.4.3            # via -r requirements/test.txt, fast-bitrix24
attrs==22.1.0             # via -r requirements/test.txt, aiohttp, pytest
beartype==0.11.0          # via -r requirements/test.txt, fast-bitrix24
bleach==5.0.1             # via -r requirements/test.txt, django-summernote, readme-renderer
certifi==2022.9.24        # via requests
cffi==1.15.1              # via cryptography
charset-normalizer==2.1.1  # via -r requirements/test.txt, aiohttp, requests
click-log==0.4.0          # via edx-lint
click==8.1.3              # via -r requirements/test.txt, click-log, code-annotations, edx-lint
code-annotations==1.3.0   # via -r requirements/test.txt, edx-lint
commonmark==0.9.1         # via rich
coverage[toml]==6.5.0     # via -r requirements/test.txt, pytest-cov
cryptography==38.0.1      # via secretstorage
dill==0.3.5.1             # via pylint
django-admin-ordering==0.16.0  # via -r requirements/test.txt
django-jazzmin==2.5.0     # via -r requirements/test.txt
django-js-asset==2.0.0    # via -r requirements/test.txt, django-admin-ordering
django-model-utils==4.2.0  # via -r requirements/test.txt
django-ninja==0.19.1      # via -r requirements/test.txt
django-summernote==0.8.20.0  # via -r requirements/test.txt
django==2.2.28            # via -c https://raw.githubusercontent.com/openedx/edx-lint/4.1.1/edx_lint/files/common_constraints.txt, -r requirements/test.txt, django-jazzmin, django-js-asset, django-model-utils, django-ninja, django-summernote
docutils==0.19            # via readme-renderer
edx-lint==5.3.0           # via -r requirements/quality.in
fast-bitrix24==1.5.12     # via -r requirements/test.txt
frozenlist==1.3.1         # via -r requirements/test.txt, aiohttp, aiosignal
icontract==2.6.2          # via -r requirements/test.txt, fast-bitrix24
idna==3.4                 # via -r requirements/test.txt, requests, yarl
importlib-metadata==5.0.0  # via keyring, twine
iniconfig==1.1.1          # via -r requirements/test.txt, pytest
isort==5.10.1             # via -r requirements/quality.in, pylint
jaraco.classes==3.2.3     # via keyring
jeepney==0.8.0            # via keyring, secretstorage
jinja2==3.1.2             # via -r requirements/test.txt, code-annotations
keyring==23.9.3           # via twine
lazy-object-proxy==1.7.1  # via astroid
markupsafe==2.1.1         # via -r requirements/test.txt, jinja2
mccabe==0.7.0             # via pylint
more-itertools==8.14.0    # via -r requirements/test.txt, fast-bitrix24, jaraco.classes
multidict==6.0.2          # via -r requirements/test.txt, aiohttp, yarl
orjson==3.8.0             # via -r requirements/test.txt
packaging==21.3           # via -r requirements/test.txt, pytest
pbr==5.10.0               # via -r requirements/test.txt, stevedore
pkginfo==1.8.3            # via twine
platformdirs==2.5.2       # via pylint
pluggy==1.0.0             # via -r requirements/test.txt, pytest
py==1.11.0                # via -r requirements/test.txt, pytest
pycodestyle==2.9.1        # via -r requirements/quality.in
pycparser==2.21           # via cffi
pydantic==1.10.2          # via -r requirements/test.txt, django-ninja
pydocstyle==6.1.1         # via -r requirements/quality.in
pygments==2.13.0          # via readme-renderer, rich
pylint-celery==0.3        # via edx-lint
pylint-django==2.5.3      # via edx-lint
pylint-plugin-utils==0.7  # via pylint-celery, pylint-django
pylint==2.15.3            # via edx-lint, pylint-celery, pylint-django, pylint-plugin-utils
pyparsing==3.0.9          # via -r requirements/test.txt, packaging
pytest-cov==4.0.0         # via -r requirements/test.txt
pytest-django==4.5.2      # via -r requirements/test.txt
pytest==7.1.3             # via -r requirements/test.txt, pytest-cov, pytest-django
python-slugify==6.1.2     # via -r requirements/test.txt, code-annotations
pytz==2022.4              # via -r requirements/test.txt, django
pyyaml==6.0               # via -r requirements/test.txt, code-annotations
readme-renderer==37.2     # via twine
requests-toolbelt==0.9.1  # via twine
requests==2.28.1          # via requests-toolbelt, twine
rfc3986==2.0.0            # via twine
rich==12.6.0              # via twine
secretstorage==3.3.3      # via keyring
six==1.16.0               # via -r requirements/test.txt, asttokens, bleach, edx-lint
snowballstemmer==2.2.0    # via pydocstyle
sqlparse==0.4.3           # via -r requirements/test.txt, django
stevedore==4.0.0          # via -r requirements/test.txt, code-annotations
text-unidecode==1.3       # via -r requirements/test.txt, python-slugify
tomli==2.0.1              # via -r requirements/test.txt, coverage, pylint, pytest
tomlkit==0.11.5           # via pylint
tqdm==4.64.1              # via -r requirements/test.txt, fast-bitrix24
twine==4.0.1              # via -r requirements/quality.in
typing-extensions==4.3.0  # via -r requirements/test.txt, astroid, icontract, pydantic, pylint, rich
urllib3==1.26.12          # via requests, twine
webencodings==0.5.1       # via -r requirements/test.txt, bleach
wrapt==1.14.1             # via astroid
yarl==1.8.1               # via -r requirements/test.txt, aiohttp
zipp==3.8.1               # via importlib-metadata
